+++ quanlynhahang/src/main/java/poly/edu/quanlynhahang/dto/IngredientBatchCreateRequest.java
@@ -8,7 +8,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record IngredientBatchCreateRequest(
-        @NotNull @Positive Double quantity,
+        @NotNull @Positive BigDecimal quantity,
         @DecimalMin("0.00") BigDecimal unitPrice,
         Date expirationDate) {
 }
