+++ quanlynhahang/src/main/java/poly/edu/quanlynhahang/dto/IngredientResponse.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 import poly.edu.quanlynhahang.entity.Ingredient;
 
-public record IngredientResponse(Long id, String name, Double quantity, String unit, Double minStock,
+public record IngredientResponse(Long id, String name, BigDecimal quantity, String unit, BigDecimal minStock,
                                  BigDecimal unitPrice, String image, Integer shelfLifeDays) {
     public static IngredientResponse from(Ingredient ingredient) {
         return new IngredientResponse(ingredient.getId(), ingredient.getName(), ingredient.getQuantity(),
