+++ quanlynhahang/src/main/java/poly/edu/quanlynhahang/dto/IngredientUpsertRequest.java
@@ -11,7 +11,7 @@ import jakarta.validation.constraints.Size;
 public record IngredientUpsertRequest(
         @NotBlank @Size(max = 200) String name,
         @NotBlank @Size(max = 50) String unit,
-        @DecimalMin("0.0") Double minStock,
+        @DecimalMin("0.0") BigDecimal minStock,
         @DecimalMin("0.00") BigDecimal unitPrice,
         @Size(max = 500) String image,
         @Min(1) @Max(3650) Integer shelfLifeDays) {
