+++ quanlynhahang/src/main/java/poly/edu/quanlynhahang/dto/PublicProductResponse.java
@@ -12,7 +12,7 @@ public record PublicProductResponse(
         Integer id,
         String name,
         BigDecimal price,
-        Double taxRate,
+        BigDecimal taxRate,
         String image,
         String description,
         Boolean status,
